Move difficulty proptests to work::tests::prop
