+++ zebra-chain/src/work/tests/prop.rs
@@ -1,3 +1,4 @@
+use primitive_types::U256;
 use proptest::{arbitrary::any, prelude::*, test_runner::Config};
 
 use std::env;
@@ -5,7 +6,7 @@ use std::env;
 use crate::block::{self, Block};
 use crate::serialization::{ZcashDeserialize, ZcashDeserializeInto, ZcashSerialize};
 
-use super::super::*;
+use super::super::{difficulty::*, *};
 
 #[test]
 fn equihash_solution_roundtrip() {
@@ -128,4 +129,91 @@ fn equihash_prop_test_input() -> color_eyre::eyre::Result<()> {
     }
 
     Ok(())
+}
+
+proptest! {
+    /// Check that CompactDifficulty expands, and converts to work.
+    ///
+    /// Make sure the conversions don't panic, and that they compare correctly.
+   #[test]
+   fn prop_compact_expand_work(compact in any::<CompactDifficulty>()) {
+       // TODO: use random ExpandedDifficulties, once we have ExpandedDifficulty::to_compact()
+       //
+       // This change will increase the number of valid random work values.
+       let expanded = compact.to_expanded();
+       let work = compact.to_work();
+
+       let hash_zero = block::Hash([0; 32]);
+       let hash_max = block::Hash([0xff; 32]);
+
+       let work_zero = Work(0);
+       let work_max = Work(u128::MAX);
+
+       if let Some(expanded) = expanded {
+           prop_assert!(expanded >= hash_zero);
+           prop_assert!(expanded <= hash_max);
+       }
+
+       if let Some(work) = work {
+           prop_assert!(work > work_zero);
+           prop_assert!(work < work_max);
+       }
+   }
+
+   /// Check that a random ExpandedDifficulty compares correctly with fixed block::Hash
+   #[test]
+   fn prop_expanded_order(expanded in any::<ExpandedDifficulty>()) {
+       // TODO: round-trip test, once we have ExpandedDifficulty::to_compact()
+       let hash_zero = block::Hash([0; 32]);
+       let hash_max = block::Hash([0xff; 32]);
+
+       prop_assert!(expanded >= hash_zero);
+       prop_assert!(expanded <= hash_max);
+   }
+
+   /// Check that ExpandedDifficulty compares correctly with a random block::Hash.
+   #[test]
+   fn prop_hash_order(hash in any::<block::Hash>()) {
+       let ex_zero = ExpandedDifficulty(U256::zero());
+       let ex_one = ExpandedDifficulty(U256::one());
+       let ex_max = ExpandedDifficulty(U256::MAX);
+
+       prop_assert!(hash >= ex_zero);
+       prop_assert!(hash <= ex_max);
+       prop_assert!(hash >= ex_one || hash == ex_zero);
+   }
+
+   /// Check that a random ExpandedDifficulty and block::Hash compare correctly.
+   #[test]
+   #[allow(clippy::double_comparisons)]
+   fn prop_expanded_hash_order(expanded in any::<ExpandedDifficulty>(), hash in any::<block::Hash>()) {
+       prop_assert!(expanded < hash || expanded > hash || expanded == hash);
+   }
+
+    /// Check that the work values for two random ExpandedDifficulties add
+    /// correctly.
+   #[test]
+    fn prop_work(compact1 in any::<CompactDifficulty>(), compact2 in any::<CompactDifficulty>()) {
+        // TODO: use random ExpandedDifficulties, once we have ExpandedDifficulty::to_compact()
+        //
+        // This change will increase the number of valid random work values.
+        let work1 = compact1.to_work();
+        let work2 = compact2.to_work();
+
+        if let (Some(work1), Some(work2)) = (work1, work2) {
+            let work_limit = Work(u128::MAX/2);
+            if work1 < work_limit && work2 < work_limit {
+                let work_total = work1 + work2;
+                prop_assert!(work_total >= work1);
+                prop_assert!(work_total >= work2);
+            } else if work1 < work_limit {
+                let work_total = work1 + work1;
+                prop_assert!(work_total >= work1);
+            } else if work2 < work_limit {
+                let work_total = work2 + work2;
+                prop_assert!(work_total >= work2);
+            }
+        }
+   }
+
 }
